improvement: better security

+++ src/API/RestService/RestApi/Commands/GroupInviteCommands/DeleteGroupInviteCommand.cs
@@ -15,17 +15,20 @@ namespace RestApi.Commands.GroupInviteCommands
 	public class DeleteGroupInviteCommand : IRequest<GroupInviteId>
 	{
 		[JsonConstructor]
-		public DeleteGroupInviteCommand(GroupInviteId groupInviteId)
-			=> GroupInviteId = groupInviteId;
+		public DeleteGroupInviteCommand(GroupInviteId groupInviteId, AppUserId appUserId)
+		{
+			GroupInviteId = groupInviteId;
+			AppUserId = appUserId;
+		}
 
 		public GroupInviteId GroupInviteId { get; }
+		public AppUserId AppUserId { get; }
 	}
 	
 	public class DeleteGroupInviteCommandHandler : IRequestHandler<DeleteGroupInviteCommand, GroupInviteId>
 	{
 		private readonly IGroupInviteRepository _groupInviteRepository;
 		private readonly IUnitOfWork _unitOfWork;
-
 
 		public DeleteGroupInviteCommandHandler(IGroupInviteRepository groupInviteRepository, IUnitOfWork unitOfWork)
 			=> (_groupInviteRepository, _unitOfWork)
@@ -41,6 +44,10 @@ namespace RestApi.Commands.GroupInviteCommands
 				throw new ApiException($"Group Invite with id:{request.GroupInviteId} does not exist",
 					StatusCodes.Status404NotFound);
 
+			if(groupInvite.InvitedAppUserId != request.AppUserId && groupInvite.InvitingAppUserId != request.AppUserId)
+				throw new ApiException("User does not have access to delete group invite",
+					StatusCodes.Status403Forbidden);
+			
 			_groupInviteRepository.Delete(groupInvite);
 			try
 			{
